bugfix at dashboard, calculate and adjustment for scale

+++ database/migrations/2022_07_20_050614_create_scale_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('scale', function (Blueprint $table) {
             $table->id();
-            $table->integer('point');
+            $table->float('point',4,2);
             $table->string("status");
             $table->timestamps();
         });
